+++ lib/modules/configuration/bloc/mission_controller/mission_variables_bloc.dart
@@ -3,17 +3,17 @@ import 'package:zenith_monitor/utils/mixins/mission_variables/class_mission_vari
 import 'package:zenith_monitor/utils/mixins/mission_variables/mission_variables_exceptions.dart';
 import 'package:zenith_monitor/pipelines/mission_pipeline/mission_bloc.dart';
 
-part 'variables_state.dart';
-part 'variables_event.dart';
+part 'mission_variables_state.dart';
+part 'mission_variables_event.dart';
 
 /// This bloc has to do with the creation of current misson variables.
 /// So, it deals with user interation in mission_creation.dart widget
 /// and update the MissionVariableList object in mission pipeline.
 
-class VariablesBloc extends Bloc<VariablesEvent, VariablesState> {
-  MissionVariablesList variablesList = MissionVariablesList();
+class MissionVariablesBloc extends Bloc<VariablesEvent, VariablesState> {
+  MissionVariablesList variablesList;
   MissionBloc missionBloc;
-  VariablesBloc(this.variablesList, this.missionBloc)
+  MissionVariablesBloc(this.variablesList, this.missionBloc)
       : super(VariablesInitial(variablesList));
 
   @override
@@ -38,7 +38,7 @@ class VariablesBloc extends Bloc<VariablesEvent, VariablesState> {
     } else if (event is StartMissionEvent) {
       try {
         variablesList.addMissionName(event.missionName);
-        //missionBloc.add(SetVariablesListEvent(variablesList: variablesList));
+        missionBloc.add(SetVariablesListEvent(variablesList: variablesList));
       } on EmptyMissionNameException {
         yield MissionNameError(
             variablesList, "É necessário fornecer um nome para a missão");
