+++ lib/modules/configuration/bloc/mission_controller/mission_variables_event.dart
@@ -1,4 +1,4 @@
-part of 'variables_bloc.dart';
+part of 'mission_variables_bloc.dart';
 
 abstract class VariablesEvent {}
