+++ lib/modules/configuration/bloc/mission_controller/mission_variables_state.dart
